More accurate EA3/EA4 detection
cPanel servers don't automatically switch to EasyApache 4 when they're updated to the version the script checks. The official way to check if the server is running EA3 or EA4 is to check if the file /etc/cpanel/ea4/is_ea4 exists. EasyApache 3 actually checks if the file exists before exiting:

stat("/etc/cpanel/ea4/is_ea4", {st_mode=S_IFREG|0644, st_size=0, ...}) = 0
write(2, "EasyApache3 is not available whe"..., 60EasyApache3 is not available when EasyApache4 is in effect.